Diff needed config and package updates ...
together with small code added inside "getGroups.ts", as required by the exercise of L6.
It is a show-case of adapting an ExpressJS app to the AWS ApiGateway + Lambda.

Also, fixed a (new) issue with using the X-Ray combined with DynamoDB - instead of importing it with "import * as AWSXRay from 'aws-xray-sdk' "
it should be added like "const AWSXRay = require('aws-xray-sdk')"

Everything was deployed to AWS and tested from the local Client, worked as expected.